Moved ExternalShuffleService to the `deploy` package + other minor tweaks.
